first implementations of EMBcomm
renamed the former comm to serialinterface (because the comm is now contained in the EMBcomm) and adapted the comm as interface between software and serial interface, not more. Based on that, the c files of the library are embedded in the source now and should work, but were not tested yet!

Changes in EMBcomm